+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/orders/wooshippinglabels/networking/DTOs.kt
@@ -27,7 +27,7 @@ data class StoreOptionsDTO(
 )
 
 data class FormDataDTO(
-    @SerializedName("selected_payment_method_id") val selectedPaymentId: Int?,
+    @SerializedName("selected_payment_method_id") val selectedPaymentId: Long?,
     @SerializedName("email_receipts") val emailReceipts: Boolean = false
 )
 
@@ -42,7 +42,7 @@ data class FormMetaDTO(
 
 data class PaymentMethodDTO(
     @SerializedName("payment_method_id")
-    val paymentMethodId: Int,
+    val paymentMethodId: Long,
     @SerializedName("name")
     val name: String,
     @SerializedName("card_type")
